나누기 소수점 나오게 수정

+++ src/day2/Calculator.java
@@ -11,7 +11,7 @@ public class Calculator {
     public int multiple(int num1 , int num2){
         return num1 * num2;
     }
-    public int divide(int num1 , int num2){
+    public float divide(int num1 , float num2){
         return num1 / num2;
     }
 }
